requirements.txt
streamlit>=1.31.0
openai>=1.0.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
